// src/users/user.interface.ts
import { NextFunction, Request, Response } from "express";

export interface IUserController {
  login: (req: Request, res: Response, next: NextFunction) => void;
  register: (req: Request, res: Response, next: NextFunction) => void;
  error: (req: Request, res: Response, next: NextFunction) => void;
}